Program Event management

+++ trunk/tests/unit/core/LocalBufferTests.cpp
@@ -66,12 +66,12 @@ TEST_F(LocalBufferTests, LocalBufferReadWriteData1Test)
         vector<float2> cpuBuf(1024);
         util::initializeBuffer(cpuBuf, 1024, 0, 1000, util::RANDOM);
 
-        ASSERT_TRUE(buf1D->readData(&cpuBuf[0], cpuBuf.size()));
+        ASSERT_TRUE(buf1D->readData(&cpuBuf[0], static_cast<unsigned long>(cpuBuf.size())));
         vector<float2> result;
         result.resize(1024);
-        ASSERT_TRUE(buf1D->writeData(&result[0], result.size()));
+        ASSERT_TRUE(buf1D->writeData(&result[0], static_cast<unsigned long>(result.size())));
 
-        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, cpuBuf.size()));
+        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, static_cast<unsigned long>(cpuBuf.size())));
 
         _bufMgr->destroyBuffer(buf1D);
     }
@@ -93,12 +93,12 @@ TEST_F(LocalBufferTests, LocalBufferReadWriteData2Test)
         vector<int2> cpuBuf(78);
         util::initializeBuffer(cpuBuf, 78, 1, 1000, util::RANDOM);
 
-        ASSERT_TRUE(buf->readData(&cpuBuf[0], cpuBuf.size()));
+        ASSERT_TRUE(buf->readData(&cpuBuf[0], static_cast<unsigned long>(cpuBuf.size())));
         vector<int2> result;
         result.resize(78);
-        ASSERT_TRUE(buf->writeData(&result[0], result.size()));
+        ASSERT_TRUE(buf->writeData(&result[0], static_cast<unsigned long>(result.size())));
 
-        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, cpuBuf.size()));
+        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, static_cast<unsigned long>(cpuBuf.size())));
 
         _bufMgr->destroyBuffer(buf);
     }
@@ -122,12 +122,12 @@ TEST_F(LocalBufferTests, LocalBufferReadWriteData3Test)
         vector<double2> cpuBuf(width * height);
         util::initializeBuffer(cpuBuf, width, height, 1000, util::RANDOM);
 
-        ASSERT_TRUE(buf1D->readData(&cpuBuf[0], cpuBuf.size()));
+        ASSERT_TRUE(buf1D->readData(&cpuBuf[0], static_cast<unsigned long>(cpuBuf.size())));
         vector<double2> result;
         result.resize(width * height);
-        ASSERT_TRUE(buf1D->writeData(&result[0], result.size()));
+        ASSERT_TRUE(buf1D->writeData(&result[0], static_cast<unsigned long>(result.size())));
 
-        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, cpuBuf.size()));
+        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, static_cast<unsigned long>(cpuBuf.size())));
 
         _bufMgr->destroyBuffer(buf1D);
     }
@@ -153,9 +153,9 @@ TEST_F(LocalBufferTests, LocalBufferReadWriteData5Test)
         int defaultVal = 123321;
         ASSERT_TRUE(buf1D->readData(&cpuBuf[0], 1200, &defaultVal));
         vector<int> result(1024 * 1024);
-        ASSERT_TRUE(buf1D->writeData(&result[0], result.size()));
+        ASSERT_TRUE(buf1D->writeData(&result[0], static_cast<unsigned long>(result.size())));
 
-        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, cpuBuf.size()));
+        ASSERT_EQ(0, util::compareBuffers(cpuBuf, result, static_cast<unsigned long>(cpuBuf.size())));
 
         _bufMgr->destroyBuffer(buf1D);
     }
